@@ -1,19 +1,14 @@
 package com.example.rjq.myapplication
 
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.view.View
-import android.widget.Button
-import android.widget.EditText
-import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
 import com.example.rjq.myapplication.progress.LoadingDialog
 import com.example.rjq.myapplication.viewmodel.MainViewModel
 import kotlinx.android.synthetic.main.activity_main.*
-import java.util.*
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : BaseActivity() {
 
     private var loadingDialog: LoadingDialog? = null
     private lateinit var viewModel: MainViewModel
@@ -21,20 +16,7 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        viewModel = ViewModelProviders.of(this).get(MainViewModel::class.java)
-
-        //生产者
-        val list1: ArrayList<out TextView> = ArrayList<Button>()
-//        val textView = list1[0]
-//        list1.add(Button(this))   //报错 禁止使用add方法
-
-        //消费者
-        val list2: ArrayList<in TextView> = ArrayList<View>() //只能add TextView：Button EditText都是TextView(多态)
-        list2.add(Button(this))
-        list2.add(EditText(this))
-        list2.add(TextView(this))
-//        list2.add(View(this))   //报错
-        val button = list2[0]                   //只能得到Any类型的
+        viewModel = ViewModelProvider(this).get(MainViewModel::class.java)
     }
 
     override fun onResume() {
