@@ -1,9 +1,9 @@
 package com.example.rjq.myapplication.viewmodel
 
 import android.app.Application
-import android.arch.lifecycle.AndroidViewModel
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MutableLiveData
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import com.example.rjq.myapplication.NoticeUtils
 import com.example.rjq.myapplication.entity.User
 import com.example.rjq.myapplication.http.HttpMethods
